build(celery): revivify celery debugging port
In the context of docker containers, localhost means the same container.
In our case, the celery worker is running in another docker container
rather than container flask. This commit revivifies the celery worker
working properly for flask applicatioin.

Besides, this commit also makes the RDB host available outside the
container, so it will be more convenient for debugging.

In the meantime, lets remove variables that we do not actually use.